requirements.txt
backoff==1.8.0
ciso8601==2.2.0
greenlet==1.1.3
jsonschema==2.6.0
psycopg2-binary==2.9.3
pyodbc==4.0.34
python-dateutil==2.8.2
pytz==2022.4
simplejson==3.11.1
singer-python==5.12.2
six==1.16.0
SQLAlchemy==1.4.41
